schema: state_or_local_gov is now 'missing', 'y' or 'n', instead of boolean. This is the first part of a larger schema cleanup.
Many attributes in the Senate XML documents can be either blank (e.g.,
'ClientState=""') or missing altogether. In the current schema, blank
values and missing attributes are both represented as NULL in the
database (None in Python). It might be useful to distinguish these two
cases.

Also, NULLs in certain columns (e.g., client.state) makes searching
for existing records a lot more complicated than it would be without
NULLs.

Therefore, this branch is all about eliminating NULLs in the schema
(and None's in Python) where there are ambiguities in the schema, or
when searching can be simplified.