lobbyists.py: DOM element parsers now return a key-value pair.
Previously they returned a list with a single key-value pair because
of an implementation detail in parse_filing. The new behavior is a bit
cleaner in case anyone wants to use just the element parsers but not
the whole module.

Also added a brief comment on what's expected of element parsers.