filings.sql: build an index for the lobbyist table.
The lobbyist table is looked up during import. Despite the fact that
inserting into the table is probably more expensive with the index, it
more than makes up for it during import in select performance. In my
tests, the elapsed time for importing the 2000_1.xml document went
from about 460s to about 70s.