setuptools packaging.
Lots of changes in this commit, but they're mostly just from moving
files around for easier setuptools support. lobbyists is now a package
and no longer a module, but all of the package symbols are visible at
in the top-level namespace, so no code changes are necessary.

Move the SQL schema into the package and tell setuptools to package it
with the source; it's effectively required in order to make use of the
importer, anyway.