Make most functions and all data structures private.
I'm about to make a release, and I want to minimize the initial public
API.

The purpose of this package is to be a reference parser and importer
for Senate LD-1/LD-2 documents, so it'd be nice if it were easily
extensible -- more of a toolkit than a production
parser/importer. That's why all functions and data structures were
initially part of the public API. However, in its current state, you
need a lot of knowledge about the internals of the parser and database
schema to make changes that don't break other parts, so hide all of
the guts for now until I come up with a more robust mechanism for
extending and/or hooking it.