Move filing/client mapping to its own table; refactor relationship insertion.
Prior to this change, filing/entity relationships were inserted by the
filing importer. Now those relationships are inserted by the
individual the entity importers instead. The problem with the previous
mechanism was that only the entity database keys are easily available
to the filing importer, not the entity fields. But in the new schema,
I want to move some of the entity fields out of the entity tables and
into the filing/entity relationship tables (e.g., client status, which
is a function of the filing, not the client entity). Making the entity
importer functions responsible for inserting their filing
relationships means it's trivial to insert any entity columns we want
into the relationship tables.